Set list of default LaTeX packages in org-mode
The list of LaTeX included packages is now redefined in the
configuration file to use an external package `org-pkgs'.

This gives some flexibility in LaTeX configuration/exporting, without
the need to modify emacs configuration file.

Due to the key position of `showlabels' package, I'm disabling visual-fill-mode,
otherwise the keys can not be seen in the buffer.