Clean slate
This is supposed to be a second "First commit" in the repo.

All the previous code and history are being kept for historical
purposes, so I'm able to find some code I wrote in the past. However,
this configuration reflects the state of my configuration in the last
couple of months, so I'm promoting this as the new stable iteration of
my config.

The major change is that, once for all, I remove the literate code from
my config. Too much effort to keeping the files for a negligible gain
and organization.